// src/enums/Genre.ts
export enum Genre {
  Fiction = "Fiction",
  LiteraryFiction = "LiteraryFiction",
  HistoricalFiction = "HistoricalFiction",
  PhilosophicalFiction = "PhilosophicalFiction",
  ScienceFiction = "ScienceFiction",
  SpeculativeFiction = "SpeculativeFiction",
  Dystopian = "Dystopian",
  Utopian = "Utopian",
  Fantasy = "Fantasy",
  MagicalRealism = "MagicalRealism",
  Horror = "Horror",
  Gothic = "Gothic",
  Mystery = "Mystery",
  Thriller = "Thriller",
  Crime = "Crime",
  Adventure = "Adventure",
  Romance = "Romance",
  Satire = "Satire",
  Comedy = "Comedy",
  Tragedy = "Tragedy",
  ComingOfAge = "ComingOfAge",
  Novella = "Novella",
  ShortStory = "ShortStory",
  NonFiction = "NonFiction",
  Biography = "Biography",
  Memoir = "Memoir",
  Essay = "Essay",
  Autobiography = "Autobiography",
  PoliticalPhilosophy = "PoliticalPhilosophy",
  Psychology = "Psychology",
  Sociology = "Sociology",
  History = "History",
  Religion = "Religion",
  Spirituality = "Spirituality",
  SelfHelp = "SelfHelp",
  Travel = "Travel",
  Science = "Science",
  Economics = "Economics",
  Philosophy = "Philosophy",
  Poetry = "Poetry",
  EpicPoetry = "EpicPoetry",
  LyricPoetry = "LyricPoetry",
  DramaticPoetry = "DramaticPoetry",
  Play = "Play",
  Drama = "Drama",
  Classic = "Classic",
  Mythology = "Mythology",
  Folklore = "Folklore",
  Renaissance = "Renaissance",
  Modernism = "Modernism",
  Postmodernism = "Postmodernism",
  Existentialism = "Existentialism",
  Absurdism = "Absurdism",
  Romanticism = "Romanticism",
  Realism = "Realism",
  Naturalism = "Naturalism",
  Surrealism = "Surrealism",
  Childrens = "Childrens",
  YoungAdult = "YoungAdult",
  FairyTale = "FairyTale",
  Fable = "Fable",
  Allegory = "Allegory",
  Anthology = "Anthology",
  GraphicNovel = "GraphicNovel",
  Experimental = "Experimental",
  Colonialism = "Colonialism",
  PoliticalFiction = "PoliticalFiction",
  LegalDrama = "LegalDrama",
  Bildungsroman = "Bildungsroman"
}